feat: move menu group package
