+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
 import kitchenpos.application.dto.MenuGroupRequest;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRepository;
+import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
