+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
 import kitchenpos.application.dto.MenuGroupRequest;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRepository;
+import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroupRepository;
 import kitchenpos.support.ServiceTest;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
